requirements.txt
paramiko
scapy
pyasn1
pyasn1_ldap
dnspython